Times series of the preprocessed data
Each zip file contains the time series preprocessed data of the subjects for the main test (49 subjects) in addition to retest (14 subjects). For the MCI subjects, since no retest were done, there are 49 subjects for the main test. Each math file is of order 130*268 such that each column describes the time series for every ROI (for the whole 268 ROIs) of every subject.